V0/requirement.txt
langchain
langchain_community
langchain_openai
langchain-pymupdf4llm
langchain_ollama
langchain_chroma
rank_bm25
python-dotenv
ipykernel
docling
langchain_experimental
streamlit
requests
pandas
fastapi[standard]
asyncio
langchain_google_genai
langchain_groq
numpy


